feat: SearchBar 폴더 생성 및 아이콘 추가
